# carwash-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Wash Admin API Documentation",
        "description": "API documentation for Car Wash Admin backend application (Laravel/Sanctum)",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login (phone/password) - Hozircha faqat telefon raqami tekshiriladi",
                "operationId": "90f464eed97fed0903a0c05651730f52",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998901234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret_password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli kirish",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                }
            }
        },
        "/v1/otp/send": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send OTP code to user via Telegram",
                "operationId": "0301b72f3aada64fd5c7a4b1e17d9aa9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998901234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP code successfully sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Verification code has been sent via Telegram."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Foydalanuvchi topilmadi"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                }
            }
        },
        "/v1/otp/verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify OTP code and issue a Sanctum token",
                "operationId": "47ca48306b418cc54d08f752592798cd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "code"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998901234567"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Noto'g'ri kod yoki muddati o'tgan"
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout the authenticated user",
                "operationId": "d30e1f479930878c138fb3d39d821260",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserResource": {
                "title": "User Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Alijon Valiyev"
                    },
                    "phone": {
                        "type": "string",
                        "example": "998901234567"
                    },
                    "email": {
                        "type": "string",
                        "example": "ali@example.com"
                    },
                    "tenantId": {
                        "type": "integer",
                        "example": 1
                    },
                    "role": {
                        "type": "string",
                        "example": "manager"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter Sanctum token for authentication (Bearer Token)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}